Make the logic available to both general and upgrade notice
